SOV-5255: update bob spice view
