+++ apps/frontend/src/app/3_organisms/Header/Header.tsx
@@ -88,12 +88,6 @@ export const Header: FC = () => {
                   className="bg-[#24BFB74D]/[0.3] border-[#24BFB74D]/[0.3] hover:bg-[#24BFB74D]"
                   onClick={() => navigate('/runes')}
                 />
-                <Button
-                  text={t(translations.header.nav.spicePoints)}
-                  style={ButtonStyle.primary}
-                  className="bg-[#24BFB74D]/[0.3] border-[#24BFB74D]/[0.3] hover:bg-[#24BFB74D] lg:invisible xl:visible"
-                  onClick={() => navigate('/bob-lp-points')}
-                />
               </>
             )}
           </ol>
