--- apps/frontend/src/app/5_pages/LeaderboardPointsPage/components/LeaderboardPoints/LeaderboardPoints.utils.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-
-import { Paragraph, prettyTx } from '@sovryn/ui';
-
-export const generateRowTitle = (id: number, wallet: string) => (
-  <Paragraph className="flex justify-between w-full">
-    <span>{id}</span>
-    <span>{prettyTx(wallet, 4)}</span>
-  </Paragraph>
-);
