+++ apps/frontend/src/locales/en/translations.json
@@ -960,7 +960,7 @@
         "title": "Spice Leaderboard",
         "description1": "Spice Points are rewards earned by providing liquidity and staking SOV on the Sovryn DEX on the BOB chain. These points track user contributions and engagement, helping determine future rewards allocated by the BOB team. This leaderboard displays only the rewards distributed by Sovryn and allows you to check your accumulated points and current position.",
         "description2": "If you want to earn your Spice points simply bridge your funds via <0>BOB Fusion and</0> start providing liquidity via our the <1>Market Making</1>  page or start <2>Staking SOV</2> tokens.",
-        "claimDescription": "Users who participated in the Season 1 Liquidity Provisioning campaign can claim their LP tokens here:",
+        "claimDescription": "Users who participated in the BOB Fusion Season 1 Spice Liquidity Provisioning campaign can claim their LP tokens here:",
         "leaderboard": "Leaderboard",
         "claimWarning": "Please note that users claiming from the distributor contract prior to <0>2024-05-31 12:00:00 UTC</0> will not be eligible to Extra Spice, and claiming prior to <1>2024-06-30 12:00:00 UTC</1> will not be eligible to Rune points.",
         "aboutCampaign": "About Extra Spice and Runes campaign",
